pyproject.toml
[project]
name = "ai-multi-agent"
version = "0.1.0"
description = "Multi-agent AI system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.2.0",
    "google-api-core>=2.0.0",
    "openai>=1.0.0",
    "langgraph>=0.0.10",
    "streamlit>=1.30.0",
    "motor>=3.3.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=24.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/govuk_content_agents"]

